Initial "entry:*" command functionality
This adds a set of commands for adding entries to the CURRENT changelog:

- entry:added
- entry:changed
- entry:deprecated
- entry:removed
- entry:fixed

Each expects the entry to add. If a --pr option is provided, and it is
an integer, the command will prepend a markdown link to the given pull
request number on GitHub. If a --package option is provided, that
package will be used to create the link; otherwise, the tool attempts to
discover it from the composer.json file.

Any multi-line notes are updated to prepend two spaces to each line.