Provide tests for ConfigCommand, and fix failing tests
This patch required doing some refactoring. While the ConfigCommand
tests passed quickly, I discovered that the `ReleaseCommand` and
`EntryCommand` needed refactoring to use values from config files when
discovered. As such, I renamed the `Provider\GetProviderTrait` to
`GetConfigValuesTrait`, and added the methods `prepareConfig()` and
`getToken()`, while modifying the signature of `getProvider()` to accept
a `Config` instance; additionally, the trait now composes
`ConfigFileTrait`, as it uses the `getConfig()` and `getConfigFile()`
methods from that trait.

These changes allow preparing a `Config` object that is initially
populated from the appropriate config file, and then overridden based on
input options. The config file is then queried when a token or provider
type is needed.