Ensure EntryCommand works!
After the previous round of changes, the `EntryCommand` would raise
errors due to attempting to retrieve the `global` and `token` options,
which were not defined in the command.

The `token` option is unneeded, so this patch changes
`GetConfigValuesTrait` to test if the token option is present before
trying to retrieve it.

The `global` option is still relevant, however, as it details where to
get the `provider` from. This option was added to the `EntryCommand`.