feat: show changelogs for specific versions
This patch adds a new command, `keep-a-changelog show-version <version>`.
When invoked, it will show the changelog for that version, along with
the date it was released.

To accomplish this, I added the method `ChangelogParser::findReleaseDateForVersion()`,
which will isolate and return the date for the given release version, if
found. Additionally, I changed the regex used when finding the date to
also accept the string `TBD`, to allow showing and fetching changelogs
for unreleased versions.

The command itself is defined in `ShowVersionsCommand`.