fix: allow GitLab subgroups in package names
This patch moves validation of the package name into the individual
providers, and expands what GitLab allows to allow subgroups in the
package name.

The various commands have been updated to indicate they allow GitLab
subgroups in package names when using GitLab and providing a package
name. Additionally, the logic in the `ReleaseCommand::createReleaseName`
method was re-worked to split the package name on the last `/` provided
(instead of using list operations with `explode`).