feat: allow providing custom domain name for provider
GitLab allows users to self-host or set a CNAME for the installation. As
such, the domain name can vary from just `gitlab.com`.

This patch adds the ability to specify the domain name in relevant
commands via:

- The `--provider-domain` option
- The `domain` configuration key in the `.keep-a-changelog.ini` config file

When provided, that domain will be used for determining the git remote
to which to push tags, as well as for creating URLs for creating
releases and changelog merge request links.