feat: create ReleaseCommand input validation event
This patch creates the `ValidateRequirementsEvent`, which will be used
to ensure that all necessary requirements (existing tag, configured
provider and token) are present.

Three listeners operate on the event:

- `VerifyTagExistsListener` verifies that the tag is actually present in
  the git repository
- `MarshalConfigurationListener` marshals the provider configuration
  from a combination of input and configuration files, using the existing
  `GetConfigValuesTrait`. No tests were written for it yet, as they
  already exist(?) in other test cases; eventually, those will be tested
  separately.
- `ValidateTokenExistsListener` ensures that the `Config` instance has a
  non-empty token present.

This patch also establishes the `ReleaseEventsProvider`, which will be
passed to an `EventDispatcher` instance in the `ReleaseCommand`. This
will enable consolidating the various input marshaling tasks to:

```php
$validator = $dispatcher->dispatch(new Release\ValidateRequirementsEvent($input, $output));
if (! $validator->requirementsMet()) {
    return 1;
}
```

Each event will compose the new `IOTrait`, which defines properties and
accessors for the symfony/console input and output instances.