feat: adds PrepareChangelogEvent and related listeners
Adds the PrepareChangelogEvent, which has the following listeners:

- DiscoverChangelogFileListener, which attempts to discover the
  changelog based on the input option, using the CHANGELOG.md in the
  current working directory if not provided. If the discovered file is
  unreadable, it notifies the event, which will halt it.
- ParseChangelogListener, which uses the ChangelogParser to find the
  changelog for the version provided to the event.
- FormatChangelogListener, which uses the ChangelogFormatter to format
  the changelog for use in a tag message.

The ReleaseEventsProvider was updated to add these listeners,
associating them with the new event.

Additionally, since changelog parsing can raise multiple exception
types, this patch adds the ExceptionInterface as a marker interface for
all package exceptions, and all exceptions were updated to implement it.