refactor: use events in ReleaseCommand
This patch updates the ReleaseCommand to compose an event dispatcher
that uses the `ReleaseEventsProvider`, and then trigger those events in
the expected order in order to do its work.

As a result, this patch adds a test case for the `ReleaseCommand`, which
also allows removing the `LookupRemoteTest` (as it is replaced by tests
of various release event listeners). Two tests are marked incomplete, as
they require a change to the `Config::provider()` method.