refactor: Refactors providers to better fit internal usage
Updates the `ProviderInterface` definition to read as follows:

```php
interface ProviderInterface
{
    /**
     * Listeners can use this to test if the provider has everything it needs
     * to create a new release, thus avoiding exceptions.
     */
    public function canCreateRelease() : bool;

    /**
     * Listeners can use this to test if the provider has everything it needs
     * to generate a patch link, thus avoiding exceptions.
     */
    public function canGeneratePatchLink() : bool;

    /**
     * @throws MissingPackageNameException
     * @throws MissingProviderTokenException
     */
    public function createRelease(
      string $releaseName,
      string $tagName,
      string $changelog
    ) : ?string;

    /**
     * @throws MissingPackageNameException
     */
    public function generateIssueLink(int $issueIdentifier) : string;

    /**
     * @throws MissingPackageNameException
     */
    public function generatePatchLink(int $patchIdentifier) : string;

    public function domain() : string;

    public function withPackageName(string $package) : self;
    public function withToken(string $token) : self;
    public function withUrl(string $url) : self;
}
```

This combines some of the functionality from the
IssueMarkupProviderInterface and ProviderNameProviderInterface, but
makes it more _behavior_ oriented. As such, those two interfaces were
removed.

The `GitHub` and `GitLab` implementations were updated accordingly, and
new exceptions created for the `Provider` subnamespace. Much of the
functionality from the existing MissingTagException was moved into a
`Provider`-specific implementation for use with the `GitHub` provider.

Tests were written covering everything but the actual API calls; these
are somewhat involved, as they nested calls to internals of the API
clients.