refactor: configuration
This patch makes configuration a first-class citizen.

It does so by updating the `Config` class to understand any item in the
system that can come from configuration or options, but which may not be
specific to a given command. These items include:

- changelog file name
- git remote name
- package name
- whether or not to prompt to save the provider token when provided via
  an input option
- providers
  - class name of the provider
  - URL to use, if different than default
  - token to use
- selected provider

Several different events can then be used to populate the configuration:

- ConfigDiscovery, which will create the `Config` instance with default
  values, and then search for global overrides, local overrides, and input
  options.
- PackageNameDiscovery, which only needs to be triggered if the package
  was not found in configuration or an option. It will look in the local
  composer.json or package.json for a name, and then resort to matching
  a git remote.
- RemoteNameDiscovery, which only needs to be triggered if the git
  remote was not found in configuration or an option. It will attempt to
  match a git remote based on the current provider and package.

A master `ConfigListener` will trigger the above events, and allows the
user to set flags detailing whether the package or remote name are
required. This listener expects a new `ConfigurableEventInterface`
instance, and will notify that instance of any failures that would
prevent continued operation.