refactor: new Config system
Rewrites the configuration system to make it first-class, and not a
tacked-on afterthought.

Configuration has four steps, each overriding the previous:

- Defaults as defined in the `Config` class.
- Providers and defaults as defined in the global configuration file,
  `$XDG_CONFIG_HOME/keep-a-changelog.ini`.
- Providers and defaults as defined in the local configuration file,
  `$PWD/.keep-a-changelog.ini`.
- Input options as provided by the user.

The config file has the following format:

```dosini
[defaults]
changelog_file = changelog.txt
provider = gitlab
remote = upstream
prompt_to_save_token = false

[providers]
github[class] = Phly\KeepAChangelog\Provider\GitHub
github[url] = https://github.mwop.net
github[token] = this-is-a-github-token
gitlab[class] = Phly\KeepAChangelog\Provider\GitLab
gitlab[token] = this-is-a-gitlab-token
```

The global and local formats are the same, but the local format also
allows specifying "package" under the `[defaults]` heading.

These then translate to the following options, as defined in
`CommonConfigOptionsTrait`:

- `--changelog|-c` (defined as a global option in `bin/keep-a-changelog`)
- `--package|-p`
- `--remote|-r`
- `--provider` (to specify a named provider as defined in a config file)
- `--provider-class` (to specify a class name for a provider to use)
- `--provider-url` (to specify the custom base URL to use for the
  selected provider)
- `--provider-token` (to specify the personal access/OAuth2 token to use
  with the selected provider)

Commands that use configuration (which will be all of them) can define
events that implement `ConfigurableEventInterface`, and attach the
`ConfigListener` as an initial listener for their own event(s), and
indicate whether or not the package or remote name are required when
gathering configuration.

That listener will trigger 1-3 events itself to gather configuration,
doing so first by hitting each of the four items, and then:

- If a package name is required, checking for it in the `composer.json`,
  `package.json`, or git remotes

- If a remote name is required, checking for it in the git remotes by
  matching URLs against the domain and package, and, if multiple are
  found, prompting the user to select one.

This approach should allow each of the commands to tailor themselves for
the given workflow, and use configuration seamlessly.