refactor: refactor ReleaseCommand following Config changes
This patch refactors the release command to use the new ConfigListener
as part of its workflow.

A `ReleaseEvent` was added with methods for indicating failure of any
required steps in the process, as well as populating artifacts necessary
for each step (version, tag name, changelog).

The changes meant that the neither the `ValidateRequirementsEvent` nor
`PushTagEvent` were necessary any longer, as each now only had listeners
more suited for the more general `ReleaseEvent`.

New listeners were added for triggering each of the following workflow events:

- `PrepareChangelogEvent` => `PrepareChangelogListener`
- `CreateReleaseEvent` => `CreateReleaseListener`

The `SaveTokenEvent` was removed entirely, as I think this should not be
a thing; users should either add them to their global configuration, or
use the `--provider-token` option.

This patch paves the way for the remaining commands to be rewritten
using command bus/event dispatcher systems.