refactor: lazy-load providers
When using the defaults -> global -> local -> options workflow for
configuration, it's entirely possible that a provider class may not
exist if it refers to one that is local to a given repository. As an
example, since things like personal access/OAuth2 tokens MUST NOT be
stored in a local configuration file, we will likely have tokens stored
in a global configuration file along with the provider name.

As such, this patch introduces the `ProviderSpec` class, which
aggregates information related to populating a `Provider` instance. The
class includes an `isComplete()` method, which verifies that a class
name is associated with the spec, that the class exists, and that the
class implements `ProviderInterface`. This method can be called only
when we know in a given workflow we need a provider, and can then be
used to report errors.

Additionally, the class includes the method `createProvider()` for
creating and populating the configured provider instance. It will fail
spectacularly if the class is invalid.

The `ProviderList` class is now updated to contain a set of
`ProviderSpec` instances instead of `ProviderInterface` instances, and
`set` was renamed to `add`, as the provider name will be pulled from the
provider instance itself.

The `Config` class was altered as follows:

- `provider()` now returns a `ProviderSpec` instance.
- `setProvider()` was renamed to `setProviderName()`, and now accepts a
  string provider name instead of a provider instance. If a package was
  previously set in the config instance, it will be passed to the
  provider spec as well.
- `setPackage()` injects the package name into the selected provider
  spec.

All listeners that interact with providers from the `Config` instance
were updated to reflect the changes.