refactor: remove CreateReleaseEvent and associated listener
Updates each of the `CreateReleaseNameListener` and
`PushReleaseToProviderListener` to consume a `ReleaseEvent`, and moves
the event failure and mutator methods consumed by those listeners into the `ReleaseEvent`.

Also creates an event-specific version of the `ConfigListener` in the
`Release` sub-namespace, allowing lazy-loading of the configuration
listener in that context.

Finally, moves the listener provider into the top-level namespace, so we
can re-use it across all commands.