refactor: make bump operations use event dispatcher
Bump operations now trigger a `BumpChangelogVersionEvent`, with the
following listeners:

- `Config\ConfigListener`
- `Common\IsChangelogReadableListener`
- `Bump\BumpChangelogVersionListener` (which performs the bump operation)

The `BumpChangelogVersionEvent` accepts EITHER a bump version method OR
a version, but NEVER both or neither. This allows it to be re-used with
the bump-to-version command, which uses the exact same workflow. The
`BumpChangelogVersionListener` then checks to see which one is composed,
and, if the bump method is present, it will lookup what the next version
should be accordingly.