refactor: use events for `remove` command
Extracts functionality from `ChangelogEditorTrait` into:

- `Common\DiscoverChangelogEntryListener`
- `ChangelogEditor`

and creates a class `Common\ChangelogEntry` for holding changelog
version details (index where it starts, its length, and the contents),
instead of using a generic `stdClass` instance.

Adds two new event interfaces, `VersionAwareEventInterface` and
`ChangelogEntryAwareEventInterface`; these can be used with any event
that composes a version and a changelog entry, respectively, and
includes common methods for indicating either discovery of the artifact,
or that it was invalid or could not be discovered. In the case of
`VersionAwareEventInterface`, functionality is provided via
`Common\VersionValidationTrait`, and a common listener,
`ValidateVersionListener`, will validate the version present in the
event.

Extracts the listener `PromptForConfirmationListener` from
`RemoveCommand`. It can `abort()` the event, which stops propagation
without marking it a failure.

Extracts the functionality in `Remove` to a listener,
`Remove/RemoveChangelogEntryListener`.

Updates `RemoveCommand` to compose an event dispatcher and dispatch a
`RemoveChangelogEntryEvent`.