refactor: update ShowVersionCommand to use event dispatcher
Creates a `ShowVersionEvent` that extends the `AbstractEvent` and
implements `VersionAwareEventInterface` (via
`VersionValidationTrait`).

Creates `ShowVersionListener`, which accepts the `ShowVersionEvent` and
pulls the changelog file and version from it in order to parse it for
the relevant changelog. If any exceptions occur, it raises an error with
an appropriate message via the event, stopping propagation and causing
it to fail.

The listener workflow re-uses the `ValidateVersionListener` and
`IsChangelogReadableListener` prior to parsing the changelog file.