refactor: Modify `EditCommand` to use event dispatcher
Creates the event `Edit\EditChangelogEntryEvent`, which extends the
abstract event and implements each of `VersionAwareEventInterface` and
`ChangelogEntryAwareEventInterface`. (Creates
`Common\ChangelogEntryDiscoveryTrait` to implement the latter, and
updates `RemoveChangelogEntryEvent` to use it instead of having its own
implementation.)

The event re-uses the `ConfigListener`, `ValidateVersionListener`
(through an extension that allows a `null` version), and the
`DiscoverChangelogEntryListener`. It adds the `DiscoverEditorListener`
(which will likely move to the `Common` namespace when refactoring the
`ConfigCommand`), and the `EditChangelogEntryListener` (which replaces
the former `Edit` class).

The `EditCommand` now creates and dispatches an
`EditChangelogEntryEvent`.