feat: adds config:show command
Usage: config:show [--global] [--local]

When both or neither flag are present, shows merged configuration.
Otherwise, shows the contents of the configuration file associated with
either flag.

When either --global or merged configuration is shown, it will mask the
contents of the provider token, merely detailing whether or not it is
present.