feat: allow prepending an issue link to entries
Adds the option `--issue|-i` for prepending an issue link to an entry to
insert into a changelog.

Adds issue identifier awareness to the `AddChangelogEntryEvent`, as well
as methods for reporting invalid issue numbers or links.

Extracts `AbstractPrependLinkListener` from `PrependPatchLinkListener`,
in order to separate out the common elements to prepending links and
allow conditional execution.