fix: get_class versus gettype in ListenerProvider
I swear I tested a few months ago and discovered that gettype would
return the class name of an object. However, it does not; it just
returns `object` for any objects. Which meant that no listeners were
ever returned by the ListenerProvider.

This patch modifies it to use `get_class()` instead.