feat: adds EditChangelogLinksListener
Creates a temp file from the discovered links (or an empty string if
none were discovered), and spawns an editor so the user can edit them.

On completion, it then either updates or appends the changelog file
(updates if links were found, appends otherwise) with the results of
editing.

If an error occurs once editing is complete, it will notify the event of
the error; otherwise, it notifies it of completion.