feat: allow adding an "Unreleased" version to the changelog
Updates the `BumpChangelogVersionEvent` to allow specifying an
"unreleased" `bumpMethod`, which will evaluate to an `unreleased`
version to add at the top of the changelog.

Updates the `BumpCommand` to allow specifying that the `$type` is
`unreleased`.

Adds a new "command" to the console, `unreleased:create`, which will
create the new version.