feat: optionally allow creating milestone when bumping/promoting versions
Each of the `bump*` commands and the `unreleased:promote` command now
accept two new options:

- `--create-milestone`
- `--create-milestone-with-name`

When the first is present, it will use either the version to which the
project is bumped OR the version specified by the user (e.g.
`bump:to-version` and `unreleased:promote`) in order to create a
milestone using the provider. This is done by triggering a
`CreateMilestoneEvent` using the provided version.

When the second option is present, it will use the value provided to it
as the new milestone name.