feat: Allow omitting version when tagging and releasing
Updates the `version:tag` and `version:release` commands to make their
`version` arguments optional. When not present, a new listener will try
and locate the latest release with a date associated, and present a
confirmation prompt to the user asking if that's the one they want to
use. Since later listeners will fail hard if the tag and/or release
already exists, this acts as a sanity check in most instances.

Fixes #84