fix: fixes test assumptions
Prophesize returns null if no rules have been setup for a method.
This works fine in many cases, but if the return value is not nullable, this will raise a type error in PHP 8 and above.
As such, we need to specifically define such rules when the return value is non-nullable.