fix: ensure release targets work
Discovered when releasing 2.12.0: In 3.4 of knplabs/github-api, the `Client::AUTH_*` constants were deprecated in favor of `AuthMethod::*` constants; in later versions in the 3.x release, the original constants were **removed** (a BC break).

This patch updates to use the `AuthMethod` constants, and pins to 3.4 and later to ensure they work.